Change codec channels from uint16 to uint8

+++ pkg/core/codec.go
@@ -13,7 +13,7 @@ import (
 type Codec struct {
 	Name        string // H264, PCMU, PCMA, opus...
 	ClockRate   uint32 // 90000, 8000, 16000...
-	Channels    uint16 // 0, 1, 2
+	Channels    uint8  // 0, 1, 2
 	FmtpLine    string
 	PayloadType uint8
 }
